Split out queries module
